jsonify inventory
